Fix incorrect vector operations
Fix problem with reward int size

+++ Task.cpp
@@ -10,7 +10,7 @@ bool stateIsTerminal(const ArmState &state) {
     return false;
 }
 
-int reward(const ArmState &state, const ArmAction action, const ArmState &statePrime) {
+int8_t reward(const ArmState &state, const ArmAction action, const ArmState &statePrime) {
     if (stateIsTerminal(statePrime)) {
         return 50;
     }
