+++ Task.h
@@ -3,5 +3,5 @@
 
 bool stateIsTerminal(const ArmState &state);
 
-int reward(const ArmState &state, const ArmAction action, const ArmState &statePrime);
+int8_t reward(const ArmState &state, const ArmAction action, const ArmState &statePrime);
 #endif
